cleanup readme
