+++ core/src/androidMain/kotlin/com/powersync/DatabaseDriverFactory.android.kt
@@ -1,31 +1,20 @@
 package com.powersync
 
 import android.content.Context
-import android.os.Build
-import androidx.annotation.RequiresApi
 import com.powersync.db.internal.InternalSchema
 import kotlinx.coroutines.CoroutineScope
 import org.sqlite.SQLiteCommitListener
-import java.util.Properties
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING", "SqlNoDataSourceInspection")
 public actual class DatabaseDriverFactory(
     private val context: Context,
 ) {
-    @RequiresApi(Build.VERSION_CODES.O)
     internal actual fun createDriver(
         scope: CoroutineScope,
         dbFilename: String,
         dbDirectory: String?,
     ): PsSqlDriver {
         val schema = InternalSchema
-
-        // WAL Mode properties
-        val properties = Properties()
-        properties.setProperty("journal_mode", "WAL")
-        properties.setProperty("journal_size_limit", "${6 * 1024 * 1024}")
-        properties.setProperty("busy_timeout", "30000")
-        properties.setProperty("cache_size", "${50 * 1024}")
 
         val dbPath =
             if (dbDirectory != null) {
@@ -38,7 +27,7 @@ public actual class DatabaseDriverFactory(
         val driver =
             JdbcSqliteDriver(
                 url = "jdbc:sqlite:$dbPath",
-                properties = properties,
+                properties = buildDefaultWalProperties(),
             )
 
         migrateDriver(driver, schema)
