+++ core/src/jvmMain/kotlin/com/powersync/DatabaseDriverFactory.jvm.kt
@@ -3,7 +3,6 @@ package com.powersync
 import com.powersync.db.internal.InternalSchema
 import kotlinx.coroutines.CoroutineScope
 import org.sqlite.SQLiteCommitListener
-import java.util.Properties
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING", "SqlNoDataSourceInspection")
 public actual class DatabaseDriverFactory {
@@ -13,13 +12,6 @@ public actual class DatabaseDriverFactory {
         dbDirectory: String?,
     ): PsSqlDriver {
         val schema = InternalSchema
-
-        // WAL Mode properties
-        val properties = Properties()
-        properties.setProperty("journal_mode", "WAL")
-        properties.setProperty("journal_size_limit", "${6 * 1024 * 1024}")
-        properties.setProperty("busy_timeout", "30000")
-        properties.setProperty("cache_size", "${50 * 1024}")
 
         val dbPath =
             if (dbDirectory != null) {
@@ -31,7 +23,7 @@ public actual class DatabaseDriverFactory {
         val driver =
             JdbcSqliteDriver(
                 url = "jdbc:sqlite:$dbPath",
-                properties = properties,
+                properties = buildDefaultWalProperties(),
             )
 
         migrateDriver(driver, schema)
